+++ index.html
@@ -64,7 +64,7 @@
         <p>Microverse Students own this website</p>
     </footer>
 
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
     
 </body>
 </html>
