+++ modules/addBook.js
@@ -1,0 +1,9 @@
+export const addBook = document.querySelector('#n2').addEventListener('click', () => {
+    document.getElementById('show-book').style.cssText = 'display: none;';
+    document.getElementById('book-form').style.cssText = 'display: flex;';
+    document.getElementById('show-contact').style.cssText = 'display: none;';
+  });
+
+
+
+
